Only run an Adapter for a Process if the previous Adapter wasn’t able to process the set
So, Adapters are put in a queue for the Processor. The processor iterates over the adapters, processing the set with each one in turn. That processing will then iterate over the instructions, processing each one of those in turn.

Previously, every Adapter was run with every instruction. Now, if the first Adapter can handle the instruction, no later adapters will try to process that instruction. If the first Adapter _cannot_ handle the instruction, it will bail and the next Adapter can try.